+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.3.116 (27 July 2025)
 
 - Added `display_name` property to `Task` (e.g. for plotting).
 - Analysis: `task_info()` operation for data frame preparation.